refactor: pointerRecorder done

+++ client/src/services/updateDateTime.js
@@ -2,7 +2,7 @@
 
 export function updateDateTime(setDate, setHour) {
     const currentDate = new Date();
-    const options = { weekday: 'long', year: 'numeric', month: 'long', day: 'numeric' };
+    const options = { weekday: 'long', year: 'numeric', month: 'numeric', day: 'numeric' };
     const formattedDate = currentDate.toLocaleDateString('pt-BR', { ...options, hour: undefined, minute: undefined, second: undefined });
 
     const timeDate = new Date();
